feat(ui): add support for keyboard shortcuts in tooltips
